added strained binaries example, and added casts to float in Binary methods

+++ src/openbandparams/iii_v/binary.py
@@ -44,7 +44,7 @@ class Binary(Base):
         '''
         Returns the lattice parameter, a, in Angstroms at 300 K.
         '''
-        return cls._a_300K
+        return float(cls._a_300K)
     
     @classmethod
     def da_dT(cls, **kwargs):
@@ -52,7 +52,7 @@ class Binary(Base):
         Returns the thermal expansion coefficient, da_dT, of the lattice
         parameter, a, in Angstroms per Kelvin.
         '''
-        return cls._da_dT
+        return float(cls._da_dT)
     
     @classmethod
     def a(cls, **kwargs):
@@ -61,7 +61,7 @@ class Binary(Base):
         temperature, T, in Kelvin (default: 300 K).
         '''
         T = cls._get_T(kwargs)
-        return cls.a_300K() + cls.da_dT() * (T - 300)
+        return float(cls.a_300K() + cls.da_dT() * (T - 300))
     
     @classmethod
     def Eg_Gamma_0(cls, **kwargs):
@@ -69,7 +69,7 @@ class Binary(Base):
         Returns the Gamma-valley bandgap, Eg_Gamma, in electron Volts
         at 0 K.
         '''
-        return cls._Eg_Gamma_0
+        return float(cls._Eg_Gamma_0)
     
     @classmethod
     def alpha_Gamma(cls, **kwargs):
@@ -77,7 +77,7 @@ class Binary(Base):
         Returns the Gamma-valley alpha Varshni parameter, alpha_Gamma, in
         electron Volts per Kelvin.
         '''
-        return cls._alpha_Gamma
+        return float(cls._alpha_Gamma)
     
     @classmethod
     def beta_Gamma(cls, **kwargs):
@@ -85,7 +85,7 @@ class Binary(Base):
         Returns the Gamma-valley beta Varshni parameter, beta_Gamma, in
         Kelvin.
         '''
-        return cls._beta_Gamma
+        return float(cls._beta_Gamma)
     
     @classmethod
     def Eg_Gamma(cls, **kwargs):
@@ -94,8 +94,8 @@ class Binary(Base):
         given temperature, T, in Kelvin (default: 300 K).
         '''
         T = cls._get_T(kwargs)
-        return varshni(cls.Eg_Gamma_0(), cls.alpha_Gamma(),
-                       cls.beta_Gamma(), T)
+        return float(varshni(cls.Eg_Gamma_0(), cls.alpha_Gamma(),
+                             cls.beta_Gamma(), T))
     
     @classmethod
     def Eg_X_0(cls, **kwargs):
@@ -103,7 +103,7 @@ class Binary(Base):
         Returns the X-valley bandgap, Eg_X, in electron Volts
         at 0 K.
         '''
-        return cls._Eg_X_0
+        return float(cls._Eg_X_0)
     
     @classmethod
     def alpha_X(cls, **kwargs):
@@ -111,7 +111,7 @@ class Binary(Base):
         Returns the X-valley alpha Varshni parameter, alpha_X, in
         electron Volts per Kelvin.
         '''
-        return cls._alpha_X
+        return float(cls._alpha_X)
     
     @classmethod
     def beta_X(cls, **kwargs):
@@ -119,7 +119,7 @@ class Binary(Base):
         Returns the X-valley beta Varshni parameter, beta_X, in
         Kelvin.
         '''
-        return cls._beta_X
+        return float(cls._beta_X)
     
     @classmethod
     def Eg_X(cls, **kwargs):
@@ -128,7 +128,8 @@ class Binary(Base):
         temperature, T, in Kelvin (default: 300 K).
         '''
         T = cls._get_T(kwargs)
-        return varshni(cls.Eg_X_0(), cls.alpha_X(), cls.beta_X(), T)
+        return float(varshni(cls.Eg_X_0(), cls.alpha_X(),
+                             cls.beta_X(), T))
     
     @classmethod
     def Eg_L_0(cls, **kwargs):
@@ -136,7 +137,7 @@ class Binary(Base):
         Returns the L-valley bandgap, Eg_L, in electron Volts
         at 0 K.
         '''
-        return cls._Eg_L_0
+        return float(cls._Eg_L_0)
     
     @classmethod
     def alpha_L(cls, **kwargs):
@@ -144,7 +145,7 @@ class Binary(Base):
         Returns the L-valley alpha Varshni parameter, alpha_L, in
         electron Volts per Kelvin.
         '''
-        return cls._alpha_L
+        return float(cls._alpha_L)
     
     @classmethod
     def beta_L(cls, **kwargs):
@@ -152,7 +153,7 @@ class Binary(Base):
         Returns the L-valley beta Varshni parameter, beta_L, in
         Kelvin.
         '''
-        return cls._beta_L
+        return float(cls._beta_L)
     
     @classmethod
     def Eg_L(cls, **kwargs):
@@ -161,7 +162,8 @@ class Binary(Base):
         temperature, T, in Kelvin (default: 300 K).
         '''
         T = cls._get_T(kwargs)
-        return varshni(cls.Eg_L_0(), cls.alpha_L(), cls.beta_L(), T)
+        return float(varshni(cls.Eg_L_0(), cls.alpha_L(),
+                             cls.beta_L(), T))
     
     @classmethod
     def Eg(cls, **kwargs):
@@ -170,14 +172,15 @@ class Binary(Base):
         temperature, T, in Kelvin (default: 300 K).
         '''
         T = cls._get_T(kwargs)
-        return min(cls.Eg_Gamma(T=T), cls.Eg_X(T=T), cls.Eg_L(T=T))
+        return float(min(cls.Eg_Gamma(T=T), cls.Eg_X(T=T),
+                         cls.Eg_L(T=T)))
     
     @classmethod
     def Delta_SO(cls, **kwargs):
         '''
         Returns the split-off energy, Delta_SO, in electron Volts.
         '''
-        return cls._Delta_SO
+        return float(cls._Delta_SO)
     
     @classmethod
     def meff_e_Gamma(cls, **kwargs):
@@ -185,7 +188,7 @@ class Binary(Base):
         Returns the Gamma-valley electron effective mass, meff_e_Gamma,
         in units of electron mass.
         '''
-        return cls._meff_e_Gamma
+        return float(cls._meff_e_Gamma)
     
     @classmethod
     def meff_e_L_long(cls, **kwargs):
@@ -193,7 +196,7 @@ class Binary(Base):
         Returns the L-valley electron effective mass in the longitudinal
         direction, meff_e_L_long, in units of electron mass.
         '''
-        return cls._meff_e_L_long
+        return float(cls._meff_e_L_long)
     
     @classmethod
     def meff_e_L_trans(cls, **kwargs):
@@ -201,7 +204,7 @@ class Binary(Base):
         Returns the L-valley electron effective mass in the transverse
         direction, meff_e_L_trans, in units of electron mass.
         '''
-        return cls._meff_e_L_trans
+        return float(cls._meff_e_L_trans)
     
     @classmethod
     def meff_e_L_DOS(cls, **kwargs):
@@ -209,7 +212,7 @@ class Binary(Base):
         Returns the L-valley electron density of states effective mass,
         meff_e_L_DOS, in units of electron mass.
         '''
-        return cls._meff_e_L_DOS
+        return float(cls._meff_e_L_DOS)
     
     @classmethod
     def meff_e_X_long(cls, **kwargs):
@@ -217,7 +220,7 @@ class Binary(Base):
         Returns the X-valley electron effective mass in the longitudinal
         direction, meff_e_X_long, in units of electron mass.
         '''
-        return cls._meff_e_X_long
+        return float(cls._meff_e_X_long)
     
     @classmethod
     def meff_e_X_trans(cls, **kwargs):
@@ -225,7 +228,7 @@ class Binary(Base):
         Returns the X-valley electron effective mass in the transverse
         direction, meff_e_X_trans, in units of electron mass.
         '''
-        return cls._meff_e_X_trans
+        return float(cls._meff_e_X_trans)
     
     @classmethod
     def meff_e_X_DOS(cls, **kwargs):
@@ -233,28 +236,28 @@ class Binary(Base):
         Returns the X-valley electron density of states effective mass,
         meff_e_X_DOS, in units of electron mass.
         '''
-        return cls._meff_e_X_DOS
+        return float(cls._meff_e_X_DOS)
     
     @classmethod
     def Luttinger1(cls, **kwargs):
         '''
         Returns the first Luttinger parameter (unitless).
         '''
-        return cls._Luttinger1
+        return float(cls._Luttinger1)
     
     @classmethod
     def Luttinger2(cls, **kwargs):
         '''
         Returns the second Luttinger parameter (unitless).
         '''
-        return cls._Luttinger2
+        return float(cls._Luttinger2)
     
     @classmethod
     def Luttinger3(cls, **kwargs):
         '''
         Returns the second Luttinger parameter (unitless).
         '''
-        return cls._Luttinger3
+        return float(cls._Luttinger3)
     
     @classmethod
     def meff_SO(cls, **kwargs):
@@ -262,21 +265,21 @@ class Binary(Base):
         Returns the split-off band hole effective mass, meff_SO, in units
         of electron mass.
         '''
-        return cls._meff_SO
+        return float(cls._meff_SO)
     
     @classmethod
     def Ep(cls, **kwargs):
         '''
         Returns the Ep matrix element, in electron Volts.
         '''
-        return cls._Ep
+        return float(cls._Ep)
     
     @classmethod
     def F(cls, **kwargs):
         '''
         Returns the F Kane parameter (unitless).
         '''
-        return cls._F
+        return float(cls._F)
     
     @classmethod
     def VBO(cls, **kwargs):
@@ -284,7 +287,7 @@ class Binary(Base):
         Returns the valance band offset energy, VBO, in electron Volts
         relative to the InSb valance band maximum.
         '''
-        return cls._VBO
+        return float(cls._VBO)
     
     @classmethod
     def a_c(cls, **kwargs):
@@ -292,7 +295,7 @@ class Binary(Base):
         Returns the conduction band deformation potential, a_c, in
         electron Volts.
         '''
-        return cls._a_c
+        return float(cls._a_c)
     
     @classmethod
     def a_v(cls, **kwargs):
@@ -300,39 +303,39 @@ class Binary(Base):
         Returns the valance band deformation potential, a_v, in
         electron Volts.
         '''
-        return cls._a_v
+        return float(cls._a_v)
     
     @classmethod
     def b(cls, **kwargs):
         '''
         Returns the b shear deformation potential, in electron Volts.
         '''
-        return cls._b
+        return float(cls._b)
     
     @classmethod
     def d(cls, **kwargs):
         '''
         Returns the d shear deformation potential, in electron Volts.
         '''
-        return cls._d
+        return float(cls._d)
     
     @classmethod
     def c_11(cls, **kwargs):
         '''
         Returns the c_11 elastic constant, in gigapascals.
         '''
-        return cls._c_11
+        return float(cls._c_11)
     
     @classmethod
     def c_12(cls, **kwargs):
         '''
         Returns the c_12 elastic constant, in gigapascals.
         '''
-        return cls._c_12
+        return float(cls._c_12)
     
     @classmethod
     def c_44(cls, **kwargs):
         '''
         Returns the c_44 elastic constant, in gigapascals.
         '''
-        return cls._c_44
+        return float(cls._c_44)
